drivers: battery_v2: add bypass charging switch
Since OneUI 4.0 update, Samsung added a Battery Protect feature by stopping charging when a certain percentage is reached.
This is accomplished by adding battery_full_capacity which tricks the driver that the battery is fully charge when it reaches a certain set point.

This commit adds a sysfs attribute battery_charging_enabled and takes advantage of the Battery Protect mechanism by setting the current battery level as battery_full_capacity when charging is disabled.

Signed-off-by: Samuel Pascua <sgpascua@ngcp.ph>